Fix the footer on small screens.
On small screens the footer appeared over the top of the text at the bottom of the page. This commit adds a background colour to the footer to make sure that you can read it, and also adds a bit of a margin to the bottom of the body so that the content at the very bottom of the page can be seen.